requirements.txt
fastapi
numpy
pandas
uvicorn
pillow
pymupdf
easyocr
opencv-python-headless
torch
torchvision
